// frontend/src/api/generated/models/gitProvider.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */

export type GitProvider = typeof GitProvider[keyof typeof GitProvider];


export const GitProvider = {
  github: 'github',
  gitlab: 'gitlab',
} as const;